Refactor navigation and contact email handling; replace hardcoded email with dynamic import and update nav data structure
