+++ data/nav.ts
@@ -1,4 +1,4 @@
-import { NavItem } from "@/types/header";
+import { NavItem } from "@/types/nav";
 
 export const navItems: NavItem[] = [
   { title: "Home", scrollTo: "hero" },
